Add miniconda

+++ run_once_install-packages.sh
@@ -2,10 +2,19 @@
 
 sudo add-apt-repository ppa:aos1/diff-so-fancy
 sudo apt update
-sudo apt install bat tmux fzf git curl ranger tree trash-cli diff-so-fancy ripgrep
-sudo apt install p7zip-full p7zip-rar
-# sudo apt install i3 scrot feh imagemagick polybar pavucontrol
-# sudo apt install neovim
 
+sudo apt install -y bat tmux fzf git curl ranger tree trash-cli diff-so-fancy ripgrep
+# Install tpm, tmux plugin manager
+sudo apt install -y p7zip-full p7zip-rar
+# if GUI prompt
+	# sudo apt install -y i3 scrot feh imagemagick polybar pavucontrol
+
+# Prompt install miniconda
+
+# Get latest neovim instead
+# Auto PlugInstall
+# sudo apt install -y neovim
+
+# Local scripts
 mkdir -p ~/.local/bin
-ln -s /usr/bin/batcat ~/.local/bin/bat
+[ ! -f "~/.local/bin/bat"] && ln -s /usr/bin/batcat ~/.local/bin/bat
